기존 ArtworkTrackList 및 DotScrollBar 제거, ArtworkPage redirect 함수 ArtworkTrackList에 추가.
