--- 0.1.0/src/scss/widgets/artworkTrack.scss
@@ -1,63 +1,0 @@
-.artworkTrack {
-    width: 100%;
-
-    .category {
-        width: 100%;
-        height: auto;
-        
-        text-align: left;
-        font-weight: 700;
-        font-size: 22px;
-        line-height: 115%;
-        letter-spacing: -0.04em;
-    
-        color: #000000;
-    }
-
-    .artwork {
-            width: 100%;
-            height: calc(var(--vw, 1vw) * 25);
-            display: flex;
-            align-items: center;
-            justify-content: flex-start;
-            border-bottom: 1px solid #AEAEAE;
-            ;
-    
-            .artworkThumbnail {
-                width: calc(var(--vw, 1vw) * 20);
-                height: 100%;
-                display: flex;
-                align-items: center;
-                justify-content: flex-start;
-    
-                img {
-                    width: calc(var(--vw, 1vw) * 17);
-                    height: calc(var(--vw, 1vw) * 17);
-                    border-radius: calc(var(--vw, 1vw) * 2);
-                }
-            }
-    
-            .artworkInformation {
-                display: flex;
-                flex-direction: column;
-                justify-content: center;
-                gap: calc(var(--vw, 1vw) * 2);
-                text-align: left;
-                color: #000000;
-    
-                .artworkName {
-                    font-weight: 600;
-                    font-size: 18px;
-                    line-height: 115%;
-                    letter-spacing: -0.04em;
-                }
-    
-                .artist {
-                    font-weight: 500;
-                    font-size: 13px;
-                    line-height: 115%;
-                    letter-spacing: -0.06em;
-                }
-            }
-        }
-}
