+++ 0.1.0/src/scss/widgets/artworkTrackList.scss
@@ -2,14 +2,100 @@
     position: relative;
     margin-top: calc(var(--vw, 1vw) * 5);
     display: flex;
-    justify-content: flex-end;
-    align-items: flex-start;
+    align-items: center;
+    justify-content: flex-start;
 
-    .artworkTracksContainer {
-        position: absolute;
-        top: 0;
-        left: 0;
-        z-index: 0;
-        width: 100%;
+    .artworkTrackListContainer {
+        flex: 1 0 auto;
+
+        .artworkTrackContainer {
+            display: flex;
+            flex-direction: column;
+            align-items: center;
+            justify-content: flex-start;
+            margin-bottom: calc(var(--vw, 1vw) * 3);
+    
+            .tracksIncategory {
+                width: 100%;
+    
+                .category {
+                    width: 100%;
+                    height: auto;
+                    text-align: left;
+                    font-weight: 700;
+                    font-size: 22px;
+                    line-height: 115%;
+                    letter-spacing: -0.04em;
+                    color: #000000;
+                }
+    
+                .track {
+                    width: 100%;
+                    height: calc(var(--vw, 1vw) * 25);
+                    display: flex;
+                    align-items: center;
+                    justify-content: space-between;
+                    border-bottom: 1px solid #AEAEAE;
+    
+                    .artwork {
+                        flex: 1 0 auto;
+                        height: 100%;
+                        display: flex;
+                        align-items: center;
+                        justify-content: flex-start;
+    
+                        .artworkThumbnail {
+                            width: calc(var(--vw, 1vw) * 20);
+                            height: 100%;
+                            display: flex;
+                            align-items: center;
+                            justify-content: flex-start;
+    
+                            .imageContainer {
+                                width: calc(var(--vw, 1vw) * 17);
+                                height: calc(var(--vw, 1vw) * 17);
+                                overflow: hidden;
+                                border-radius: calc(var(--vw, 1vw) * 2);
+                                position: relative;
+    
+                                img {
+                                    position: absolute;
+                                    top: 50%;
+                                    left: 50%;
+                                    transform: translate(-50%, -50%);
+                                }
+                            }
+                        }
+    
+                        .artworkInformation {
+                            display: flex;
+                            flex-direction: column;
+                            justify-content: center;
+                            gap: calc(var(--vw, 1vw) * 2);
+                            text-align: left;
+                            color: #000000;
+    
+                            .artworkName {
+                                font-weight: 600;
+                                font-size: 18px;
+                                line-height: 115%;
+                                letter-spacing: -0.04em;
+                            }
+    
+                            .artist {
+                                font-weight: 500;
+                                font-size: 13px;
+                                line-height: 115%;
+                                letter-spacing: -0.06em;
+                            }
+                        }
+                    }
+                }
+            }
+        }
+    }
+    
+    ::v-deep(.dotScrollBar) {
+        align-self: flex-start;
     }
 }
