+++ 0.1.0/src/scss/widgets/dotScrollBar.scss
@@ -1,13 +1,13 @@
 .dotScrollBar {
     position: sticky;
-    z-index: 1;;
-    top: 0px;
+    z-index: 1;
+    top: 70px;
     width: 4px;
     border-radius: 2px;
     background-color: #AEAEAE;
     height: 0px;
     overflow: visible;
-    margin-top: calc(var(--vw, 1vw) * 10);
+    margin-top: calc(var(--vw, 1vw) * 3);
 
     .categoryButton {
         position: absolute;
@@ -15,7 +15,7 @@
         background-color: transparent;
         right: 2px;
         transform: translate(50%, 0);
-        
+
         .categoryCircle {
             position: absolute;
             z-index: 1;
@@ -47,6 +47,7 @@
             letter-spacing: -0.06em;
         }
     }
+
     .progressBarContainer {
         position: relative;
         z-index: 0;
@@ -55,7 +56,7 @@
         border-radius: 2px;
         background-color: #AEAEAE;
         overflow: hidden;
-        
+
         .progressBar {
             position: relative;
             z-index: 0;
@@ -67,5 +68,5 @@
             background-color: black;
         }
     }
-    
+
 }
