--- 0.1.0/src/scss/widtets/NewDotScrollBar.scss
@@ -1,72 +1,0 @@
-.dotScrollBar {
-    position: sticky;
-    z-index: 1;
-    top: 70px;
-    width: 4px;
-    border-radius: 2px;
-    background-color: #AEAEAE;
-    height: 0px;
-    overflow: visible;
-    margin-top: calc(var(--vw, 1vw) * 3);
-
-    .categoryButton {
-        position: absolute;
-        z-index: 1;
-        background-color: transparent;
-        right: 2px;
-        transform: translate(50%, 0);
-
-        .categoryCircle {
-            position: absolute;
-            z-index: 1;
-            opacity: 1;
-            top: 50%;
-            left: 50%;
-            transform: translate(-50%, -50%);
-            width: calc(var(--vw, 1vw) * 4.5);
-            height: calc(var(--vw, 1vw) * 4.5);
-            border-radius: 50%;
-        }
-
-        .categoryName {
-            position: absolute;
-            z-index: 0;
-            opacity: 0;
-            top: 50%;
-            left: 50%;
-            transform: translate(-80%, -50%);
-            width: calc(var(--vw, 1vw) * 14);
-            height: calc(var(--vw, 1vw) * 4.6);
-            border-radius: calc(var(--vw, 1vw) * 2.3);
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            font-weight: 500;
-            font-size: 12px;
-            line-height: 115%;
-            letter-spacing: -0.06em;
-        }
-    }
-
-    .progressBarContainer {
-        position: relative;
-        z-index: 0;
-        width: 100%;
-        height: 100%;
-        border-radius: 2px;
-        background-color: #AEAEAE;
-        overflow: hidden;
-
-        .progressBar {
-            position: relative;
-            z-index: 0;
-            width: 100%;
-            height: 100%;
-            border-radius: 2px;
-            transform-origin: 50% 0;
-            transform: scaleY(0);
-            background-color: black;
-        }
-    }
-
-}
