+++ 0.1.0/src/widgets/ArtworkTrackList.vue
@@ -1,121 +1,115 @@
 <template>
     <div class="artworkTrackList">
-        <DotScrollBar :position_array="this.position_array" 
-            :contents_height="this.track_list_height"
-            :contents_top_position="this.track_list_position"
-            :document_element_id="this.document_element_id">
-        </DotScrollBar>
-        <div class="artworkTracksContainer">
-            <ArtworkTrack class="before-enter" v-for="(track, i) in this.track_array" 
-                :key="i" :track="track" :track_number="i.toString()"
-                :track_page_id_array="this.track_page_id_array"
-                :source="this.source"
-                @track_loaded="this.getTrackListProperty()"></ArtworkTrack>
+        <div class="artworkTrackListContainer" ref="artworkTrackListContainer">
+            <div class="artworkTrackContainer" v-for="(category, i) in this.category_list">
+                <div class="tracksIncategory">
+                    <div class="category">{{ category }}</div>
+                    <div v-for="(track, j) in this.artwork_track_list[i]" :key="j" class="track" @click="this.redirectArtworkPage(i, j)">
+                        <div class="artwork">
+                            <div class="artworkThumbnail">
+                                <div class="imageContainer">
+                                    <img class="thumbnailImage" :src="track.thumbnail" :style="track.style">
+                                </div>
+                            </div>
+                            <div class="artworkInformation">
+                                <div class="artworkName poppins">{{ track.getName() }}</div>
+                                <div class=" artist">{{ track.getArtistName() }}</div>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
         </div>
+        <DotScrollBar
+            :basis_element_id="'viewPort'"
+            :scroll_basis_element_id="'body'"
+            :scroll_element_class="'artworkTrackListContainer'"
+            :category_list="this.category_list" 
+            :artwork_track_list="this.artwork_track_list"
+            :contents_height="this.contents_height"
+        ></DotScrollBar>
     </div>
 </template>
 <script>
-    import ArtworkTrack from '@/widgets/ArtworkTrack.vue'
-    import DotScrollBar from './DotScrollBar.vue';
+    import { cropImage } from '@/modules/image'
+    import DotScrollBar from './DotScrollBar.vue'
+
     export default {
         name: 'ArtworkTrackList',
-        components: {
-            ArtworkTrack,
-            DotScrollBar,
-        },
+        components: { DotScrollBar },
         props: {
-            source: {
-                type: String,
-                default: ''
+            //source: String,
+            prop_category_list: Array,
+            prop_artwork_track_list: Array,
+        },
+        watch: {
+            'prop_category_list': {
+                prop_category_list: true,
+                handler () {
+                    this.category_list = this.prop_category_list.slice()
+                    this.contents_height = this.$refs.artworkTrackListContainer.clientHeight
+                }
             },
-            document_element_id: {
-                type: String,
-                default: 'app'
+            'prop_artwork_track_list': {
+                deep: true,
+                async handler() {
+                    this.artwork_track_list = this.prop_artwork_track_list.map(v => v.slice())
+                    this.artwork_track_list.forEach(async (value) => {
+                        value.forEach(async (artwork, index) => {
+                            artwork.thumbnail = await artwork.getThumbnailImage()
+                            artwork.style = await cropImage(artwork.thumbnail, 1)
+                        })
+                    })
+                    this.contents_height = this.$refs.artworkTrackListContainer.clientHeight
+                }
             },
-            artwork_track_list: {
-                type: Array,
-                default: []
-            },
-            category_list: {
-                type: Array,
-                default: []
-            },
-            proper_position_flag: {
-                type: Boolean,
-                default: false
-            }
         },
         data() {
             return {
-                track_total_number: 0,
-                track_page_id_array: null,
-                track_list_height: null,
-                track_list_position: null,
+                category_list: null,
+                artwork_track_list: null,
+                contents_height: null
             };
         },
-        watch: {
-            proper_position_flag () {
-                this.getTrackListProperty(0)
-            }
-        },
-        computed: {
-            track_array () {
-                if (this.artwork_track_list.length === 0 || this.category_list.length === 0) {
-                    return new Array(0)
-                }
-                else {
-                    this.track_page_id_array = new Array(this.artwork_track_list.length)
-
-                    let track_array = new Array(this.artwork_track_list.length)
-
-                    for (let i = 0; i < track_array.length; i++) {
-                        let track = new Object()
-
-                        track.category = (this.category_list[i] ? this.category_list[i] : null)
-                        track.artwork = this.artwork_track_list[i]
-                        
-                        track_array[i] = track
-
-                        this.track_page_id_array[i] = this.artwork_track_list[i].page_id
-                    }
-                    return track_array
-                }
-            },
-            position_array () {
-                if (this.artwork_track_list.length === this.category_list.length) {
-                    return this.category_list
-                }
-                else {
-                    let position_array = new Array(this.artwork_track_list.length)
-
-                    for (let i = 0; i < position_array.length; i++) {
-                        position_array[i] = this.category_list[i] ? this.category_list[i] : null
-                    }
-
-                    return position_array
-                }
-            }
-        },
+        beforeCreate() {},
+        created() {},
+        beforeMount() {},
+        mounted() {},
+        beforeUpdate() {},
+        updated() {},
+        beforeUnmount() {},
+        unmounted() {},
         methods: {
-            getTrackListProperty(track_load_number = 1) {
-                // proper_position_flag 값 변화(false -> true)에 의해 호출된 경우
-                if (this.proper_position_flag && track_load_number == 0) {
-                    this.track_list_position = parseInt(document.getElementsByClassName('artworkTrack 0')[0].getBoundingClientRect().top + document.getElementById(this.document_element_id).scrollTop)
-                }
-                // artworkTrack들의 load event 수신자로써 호출된 경우
-                else {
-                    this.track_total_number += track_load_number
-            
-                    if (this.track_total_number === this.artwork_track_list.length) {
-                        this.track_list_height = parseInt(document.getElementsByClassName('artworkTracksContainer')[0].clientHeight)
-                        console.log(this.track_list_height)
+            redirectArtworkPage (category_index, artwork_index) {
+                let track_page_id_array = new Array(0)
+                let artwork_page_index = 0
+                let artwork_count = 0
+                this.artwork_track_list.forEach((value, index) => {
+                    value.forEach((artwork, index_2) => {
+                        track_page_id_array.push(artwork.getPageID())
 
-                        document.getElementsByClassName('artworkTrackList')[0].style.setProperty('height', `${document.getElementsByClassName('artworkTrackList')[0].clientHeight + this.track_list_height}px`)
-                    }
+                        if (index == category_index && index_2 == artwork_index) {
+                            artwork_page_index = artwork_count
+                        }
+                        artwork_count++
+                    })
+                })
+                if (process.env.NODE_ENV === 'production') {
+                    this.$gtag.event('click', {
+                        event_category: 'exhibition',
+                        event_label: 'enter',
+                        value: this.artwork_track_list[category_index][artwork_index].getID()
+                    })
                 }
-                
-                return
+                this.$router.push({
+                    path: '/artwork',
+                    query: {
+                        array: track_page_id_array,
+                        index: artwork_page_index
+                    }
+                })
             }
+            
         }
     }
 </script>
